refactor logic to helper class
