+++ services/api/tests/Firefly.Signal.JobSearch.FunctionalTests/Application/JobSearchQueriesTests.cs
@@ -116,7 +116,7 @@ public sealed class JobSearchQueriesTests
         Assert.AreEqual(newerRun.Id, runs.Items[0].Id);
         Assert.AreEqual("Adzuna", runs.Items[0].ProviderName);
         Assert.AreEqual(JobRefreshRunStatus.Failed.ToString(), runs.Items[0].Status);
-        Assert.AreEqual("{\"keyword\":\"Developer\",\"pageSize\":50,\"postcode\":\"E15\",\"provider\":0,\"pageIndex\":1,\"titleOnly\":false}", runs.Items[0].JsonFilter);
+        Assert.AreEqual("{\"keyword\":\"Developer\",\"pageSize\":50,\"postcode\":\"E15\",\"provider\":\"Adzuna\",\"pageIndex\":1,\"titleOnly\":false}", runs.Items[0].JsonFilter);
         Assert.AreEqual(2, runs.Items[0].RecordsReceived);
         Assert.AreEqual(0, runs.Items[0].RecordsInserted);
         Assert.AreEqual(1, runs.Items[0].RecordsFailed);
